fix(exporters/elasticsearch-exporter): migrates ES to pkg io.zeebe.exporter
