+++ exporters/elasticsearch-exporter/src/main/java/io/zeebe/exporter/ElasticsearchExporter.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.zeebe.exporter.api;
+package io.zeebe.exporter;
 
-import io.zeebe.exporter.api.ElasticsearchExporterConfiguration.IndexConfiguration;
+import io.zeebe.exporter.ElasticsearchExporterConfiguration.IndexConfiguration;
 import io.zeebe.exporter.api.context.Context;
 import io.zeebe.exporter.api.context.Controller;
 import io.zeebe.exporter.api.record.Record;
