+++ exporters/elasticsearch-exporter/src/test/java/io/zeebe/exporter/util/ElasticsearchForkedJvm.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.zeebe.exporter.api.util;
+package io.zeebe.exporter.util;
 
 import java.io.File;
 import java.io.IOException;
